moved PingExpectedImprovement, PingOnePotentialSampleExpectedImprovement and PingEITest from gpp_math_test.cpp to gpp_math_test.hpp, also impose min version of 1.51 for boost in CMakeLists.txt
