+++ moe/optimal_learning/cpp/CMakeLists.txt
@@ -53,7 +53,7 @@ endif()
 
 #### External dependencies
 find_package(PythonLibs REQUIRED)
-find_package(Boost COMPONENTS python REQUIRED)
+find_package(Boost 1.51 COMPONENTS python REQUIRED)
 
 include_directories(${PYTHON_INCLUDE_DIRS})
 include_directories(${Boost_INCLUDE_DIRS})
